requirements.txt
matplotlib
Pillow
# Optional dependency for PDF output
# cairosvg
